instruments renamed back to pairs
